migrate EntityTag related styles and style props to EntityTag
